Refactor: move all meditrak files into apiV2/meditrakApp
