+++ packages/central-server/src/apiV2/index.js
@@ -18,9 +18,7 @@ import { useRouteHandler } from './RouteHandler';
 import { exportRoutes } from './export';
 import { importRoutes } from './import';
 import { authenticate } from './authenticate';
-import { changesMetadata } from './changesMetadata';
-import { countChanges } from './countChanges';
-import { getChanges } from './getChanges';
+import { countChanges, changesMetadata, getChanges, getSocialFeed, getUserRewards, postChanges } from './meditrakApp';
 import { BESAdminCreateHandler } from './CreateHandler';
 import { BESAdminDeleteHandler } from './DeleteHandler';
 import { BESAdminEditHandler } from './EditHandler';
@@ -91,13 +89,10 @@ import {
 import { EditEntity, GETEntities, DeleteEntity } from './entities';
 import { GetEntityTypes } from './entityTypes';
 import { EditAccessRequests, GETAccessRequests } from './accessRequests';
-import { postChanges } from './postChanges';
 import { changePassword } from './changePassword';
 import { deleteAccount } from './deleteAccount';
 import { editUser } from './editUser';
 import { requestCountryAccess } from './requestCountryAccess';
-import { getSocialFeed } from './getSocialFeed';
-import { getUserRewards } from './getUserRewards';
 import { requestPasswordReset } from './requestPasswordReset';
 import { getCountryAccessList } from './getCountryAccessList';
 import { surveyResponse } from './surveyResponse';
