+++ packages/central-server/src/apiV2/meditrakApp/countChanges/LegacyCountChangesHandler.js
@@ -4,8 +4,9 @@
  */
 
 import { getKeysSortedByValues, respond, UnauthenticatedError } from '@tupaia/utils';
-import { getUniversalTypes } from '../../database/utilities';
-import { fetchRequestingMeditrakDevice, buildMeditrakSyncQuery } from '../utilities';
+import { getUniversalTypes } from '../../../database/utilities';
+import { buildMeditrakSyncQuery } from '../meditrakSync';
+import { fetchRequestingMeditrakDevice } from '../utilities';
 
 const MAX_FAILS_BEFORE_LOG_OUT = 2;
 const MAX_FAILS_BEFORE_TYPE_EXCLUSION = 5;
