+++ packages/central-server/src/apiV2/meditrakApp/countChanges/countChanges.js
@@ -4,9 +4,9 @@
  */
 
 import { respond, DatabaseError, UnauthenticatedError } from '@tupaia/utils';
-import { buildMeditrakSyncQuery } from '../utilities';
 import { LegacyCountChangesHandler } from './LegacyCountChangesHandler';
-import { allowNoPermissions } from '../../permissions';
+import { allowNoPermissions } from '../../../permissions';
+import { buildMeditrakSyncQuery } from "../meditrakSync";
 
 const handleNonLegacyRequest = async (req, res) => {
   const { models } = req;
