+++ packages/central-server/src/apiV2/meditrakApp/getUserRewards.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 import { respond } from '@tupaia/utils';
-import { getRewardsForUser } from '../social';
+import { getRewardsForUser } from '../../social';
 
 // TODO: Remove as part of RN-502
 export const getUserRewards = async (req, res) => {
