+++ packages/central-server/src/apiV2/meditrakApp/utilities/addSurveyImage.js
