+++ packages/central-server/src/dataAccessors/index.js
@@ -3,13 +3,11 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 
-export { addSurveyImage } from './addSurveyImage';
 export { editSurvey } from './editSurvey';
 export { findQuestionsInSurvey } from './findQuestionsInSurvey';
 export { findAnswersInSurveyResponse } from './findAnswersInSurveyResponse';
 export { deleteScreensForSurvey } from './deleteScreensForSurvey';
 export { deleteOrphanQuestions } from './deleteOrphanQuestions';
 export { createUser } from './createUser';
-export { updateOrCreateSurveyResponse } from './updateOrCreateSurveyResponse';
 export { validateSurveyFields } from './validateSurveyFields';
 export { upsertAnswers } from './upsertAnswers';
